kscale/requirements.txt
# requirements.txt

# Configuration
omegaconf
email_validator
colorlogging

# HTTP requests
aiohttp
cryptography
httpx
pydantic
pyjwt
requests
yarl

# CLI
aiofiles
click
tabulate

# Async
async-lru

# K-Scale
krec
